Fixing bug related to right click menu on layouts (#520)
**User-Facing Changes**
Fixed a bug on the Layouts tab where right-clicking any layout, while
another layout with unsaved changes was selected, incorrectly showed
`Save changes` and `Revert` options, which should only appear for
layouts with unsaved changes.


![image](https://github.com/user-attachments/assets/94d086e8-18bd-497f-a84e-8182b64963e8)


**Description**
This bug occurred because the condition for adding the extra menu
options relied on the `anySelectedModifiedLayouts` variable. As a
result, whenever a modified layout was selected, the menu would include
`Save changes` and `Revert`, even when right-clicking a different
layout. This has been fixed by basing the menu options on the state of
the layout that was right-clicked, rather than the selected one.

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

**Checklist**

- [X] The web version was tested and it is running ok
- [X] The desktop version was tested and it is running ok
- [X] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated