Drag and drop for Layouts (#522)
**User-Facing Changes**
To keep Lichtblick features consistent users will now be able to also
drag-and-drop Layouts, just like it's doable with Extensions and MCAPs.

**Description**
I've updated the type of file extensions allowed to also include `.json`
files. After I've changed the hook that deals with all the files that
are dropped onto Lichtblick so that when a `.json` file is dropped that
file, or files, will be saved on IndexedDB where all the layouts are
stored. For that I had to extract part of the logic that was used on
`importLayout` and create a separate function, `parseAndInstallLayout`
that could be used either on import or on drag and drop.

**Note:** Right now there isn't a logic implemented to check if the
`.json` is actually a layout, certain `.json` will give an error, but if
they follow the normal JSON syntax they'll be accepted. Checking the
file attributes isn't an option because the metadata is equal on all
JSON files. The only way I could see a way to verify if it's an actual
layout is by reading the content of the files and checking for certain
attributes only layout files have.


<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

**Checklist**

- [X] The web version was tested and it is running ok
- [X] The desktop version was tested and it is running ok
- [X] This change is covered by unit tests
- [X] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated