Feature/Refactor desktop extension handler with tests (#542)
**User-Facing Changes**
N/A
<!-- will be used as a changelog entry -->

**Description**
The `extension.ts` file that handles operations for extensions in the
Desktop version was changed to `ExtensionHandler.ts` and transformed
into a class. This makes it possible to use mock functions and create
more reliable unit tests for these operations.

In this transformation to a class, I did not focus on changing the logic
of the operations, only structural changes.

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

**Checklist**

- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok
- [x] This change is covered by unit tests
- [x] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated

---------

Signed-off-by: Luiz Bezerra <luiz.bezerra@ctw.bmwgroup.com>
Co-authored-by: Gabriela Almeida (CTW02164) <gabriela.almeida@ctw.bmwgroup.com>