Accept any image format (#577)
**User-Facing Changes**
Users can input any string to image format.

**Description**
On PR #471, during the implementation of compressed video, it was added
a list to check for supported image formats. While this is conceptually
correct, the new `format` used on ROS became incompatible. This PR
removed this check, only video is checked.

This PR resolves #575

**Checklist**

- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok
- [x] This change is covered by unit tests