Preloading TF as false by default (#620)
## User-Facing Changes
Lichtblick will start with preload TF as false by default, improving
performance. Users that rely on the transform tree may need to turn it
on again to have a more precise visualization.

## Description
Preload TF starts as false by default, it will "hide" the performance
problem. A definitive solution is already on the way when we replace the
Block Loader with the [Tree
Loader](https://github.com/lichtblick-suite/lichtblick/tree/feature/tree-loader)

This was already discussed on issue #382 in the past

It was implemented also a tooltip for all input fields. For this
specific case, a message to warn the user about performance was added:
<img width="794" height="333" alt="image"
src="https://github.com/user-attachments/assets/9f995f13-3667-4b1a-9c0f-1299e2fed1f0"
/>

## Checklist

- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok