Fix displayTemporaryError (#616)
## User-Facing Changes

When SceneExtensions try and use the Renderer's displayTemporaryError()
function, it now properly displays it on the main window's
snackbar/toast.

## Description

Previously, `displayTemporaryError()` as defined in `ThreeDeeRender.tsx`
tried to use:
```ts
const { enqueueSnackbar } = useSnackbar();
```

Calling this `enqueueSnackbar()` did nothing because the
ThreeDeeRender.tsx sits inside an isolated new sync root that doesn't
have access to the `StudioToastProvider` and `SnackbarProvider`.

This PR passes `enqueueSnackbarFromParent` in through the
ThreeDeeRender's props so that the isolated context can share the same
toast through `displayTemporaryError()`.

## Slight caveats

The studio toast does not seem to be temporary; it doesn't auto dismiss
after some time and requires manual dismissal.
I think this should be fine since you won't run into this often at all.
At this point in time, `displayTemporaryError()` is called only in
`ImageMode.ts` for when downloading the camera image fails.
I also call it in another feature I'm developing but decided to split
the fix out to this separate PR.

## Video

https://github.com/user-attachments/assets/d25d224c-10cd-4635-92ba-99aad66bb33e

## Checklist

- [x] The web version was tested and it is running ok
- [x] The desktop version was tested and it is running ok
- [x] ~This change is covered by unit tests~
- [x] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated