Implement git flow based pipeline (#645)
## User-Facing Changes

NA

## Description

Implements a GitFlow-based release workflow.
A pre-release with a version in the format vX.Y.Z-rc.N can be manually
triggered on demand.
Every merge into main triggers a full release — if it’s from a hotfix
branch, it bumps the patch version; if it’s from a release branch, it
bumps the minor version.

### Repo Settings:

1 - A new ruleset to protect branch `develop` from merge was created.
2 - The new base branch will be `develop` which will raise the need for
updating every existing PR to merge with develop.
3 - Branches with naming that is not compliant with repo rules will
trigger the failure of `enforce-branch-naming` worflow.


## Checklist

- [ ] The web version was tested and it is running ok
- [ ] The desktop version was tested and it is running ok
- [ ] This change is covered by unit tests
- [ ] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated