Fix opening files by source flag on second instance (#683)
## User-Facing Changes
Users can now open a second instance, using the flag
`--force-multiple-windows` and `--source`.

## Description
In the implementation of `--force-multiple-windows`, it was passed the
files arguments to be opened, but not the files in the `--source`. It
was merged and unified to be in one place, and called accordingly.

Unit tests were also added. 

## Checklist

- [X] The desktop version was tested and it is running ok
- [X] This change is covered by unit tests
- [X] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated